changed the way to login and logout , removed me endpointes ,used another way to resort reviews and addresses

+++ drivers/views.py
@@ -1,38 +1,12 @@
-#  Copyright (c) Code Written and Tested by Ahmed Emad in 02/01/2020, 20:19
-from django.contrib.auth import login, authenticate, logout
-from django.contrib.auth.decorators import login_required
+#  Copyright (c) Code Written and Tested by Ahmed Emad in 03/01/2020, 19:48
+from django.contrib.auth import login
 from rest_framework import viewsets, status
-from rest_framework.decorators import api_view
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
 from drivers.models import DriverProfileModel, DriverReviewModel
 from drivers.permissions import DriverProfilePermissions, DriverReviewPermissions
 from drivers.serializers import DriverProfileSerializer, DriverReviewSerializer
-
-
-@api_view(['POST'])
-def login_view(request):
-    if request.user.is_authenticated:
-        return Response('Driver already logged in', status=status.HTTP_401_UNAUTHORIZED)
-
-    username = request.data['username']
-    password = request.data['password']
-
-    user = authenticate(username=username, password=password)
-
-    if user and hasattr(user, 'driver_profile'):
-        login(request, user)
-        return Response('Logged In Successfully')
-    else:
-        return Response('Wrong Username or Password', status=status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(['POST'])
-@login_required
-def logout_view(request):
-    logout(request)
-    return Response('Logged Out Successfully')
 
 
 class DriverProfileView(viewsets.ViewSet):
@@ -66,15 +40,9 @@ class DriverProfileView(viewsets.ViewSet):
         return Response(serializer.data)
 
     def retrieve(self, request, username=None):
-        if username:
-            driver_profile = get_object_or_404(DriverProfileModel, account__username=username)
-            serializer = DriverProfileSerializer(driver_profile)
-            return Response(serializer.data)
-        if request.user.is_authenticated and hasattr(request.user, 'driver_profile'):
-            driver_profile = request.user.driver_profile
-            serializer = DriverProfileSerializer(driver_profile)
-            return Response(serializer.data)
-        return Response(status=status.HTTP_401_UNAUTHORIZED)
+        driver_profile = get_object_or_404(DriverProfileModel, account__username=username)
+        serializer = DriverProfileSerializer(driver_profile)
+        return Response(serializer.data)
 
     def create(self, request):
         if not request.user.is_authenticated:
@@ -86,24 +54,27 @@ class DriverProfileView(viewsets.ViewSet):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         return Response(status=status.HTTP_401_UNAUTHORIZED)
 
-    def update(self, request):
-        driver_profile = request.user.driver_profile
+    def update(self, request, username=None):
+        driver_profile = get_object_or_404(DriverProfileModel, account__username=username)
+        self.check_object_permissions(request, driver_profile)
         serializer = DriverProfileSerializer(driver_profile, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def partial_update(self, request):
-        driver_profile = request.user.driver_profile
+    def partial_update(self, request, username=None):
+        driver_profile = get_object_or_404(DriverProfileModel, account__username=username)
+        self.check_object_permissions(request, driver_profile)
         serializer = DriverProfileSerializer(driver_profile, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def destroy(self, request):
-        driver_profile = request.user.driver_profile
+    def destroy(self, request, username=None):
+        driver_profile = get_object_or_404(DriverProfileModel, account__username=username)
+        self.check_object_permissions(request, driver_profile)
         driver_profile.account.delete()
         driver_profile.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -155,6 +126,6 @@ class DriverReviewView(viewsets.ViewSet):
         review.delete()
 
         driver.calculate_rating()
-        driver.resort_reviews()
+        driver.resort_reviews(review.sort)
 
         return Response(status=status.HTTP_204_NO_CONTENT)
